chore(protocol): merge protocol messages
